Changed some integers to work as doubles for future changes.

+++ FtcRobotController/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/OpModeCougarWoods3.java
@@ -67,10 +67,10 @@ public class OpModeCougarWoods3 extends OpMode{
     double flt = 0;
     double brt = 0;
     double blt = 0;
-    int fro = 0;
-    int flo = 0;
-    int bro = 0;
-    int blo = 0;
+    double fro = 0;
+    double flo = 0;
+    double bro = 0;
+    double blo = 0;
 
     if ((gamepad1.left_stick_y >= 0.05) || (gamepad1.left_stick_y <= -0.05)) {
       frm += gamepad1.left_stick_y;
